Move ClassEntry and siblings to package com.oracle.objectfile.debugentry
