+++ substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/elf/dwarf/DwarfFrameSectionImpl.java
@@ -26,6 +26,8 @@
 
 package com.oracle.objectfile.elf.dwarf;
 import com.oracle.objectfile.LayoutDecision;
+import com.oracle.objectfile.debugentry.ClassEntry;
+import com.oracle.objectfile.debugentry.PrimaryEntry;
 import com.oracle.objectfile.debuginfo.DebugInfoProvider;
 
 import static com.oracle.objectfile.elf.dwarf.DwarfSections.DW_CFA_CIE_id;
